refactor(ai-agent): rename agent_registry to registry and trim lib.rs pub use surface
