+++ crates/aionui-ai-agent/src/registry.rs
